Basic twochart test (pie, bar)
Just basic test of drawing data that originates in R using dc.js

Create two linked charts, a pie and a bar.  Expect R to send down a
list of rows, each of which is a list with at least two columns.
Treat both of them as ordinal values and draw a pie chart of the
number of rows for each value of first-col, and a bar char of number
of rows for each value of second-col.