use the result of update_notebook for fork reload rather than loading
github has both write-read and write-write race conditions.
in this case, doing an update and then immediately loading the gist
was sometimes returning the old notebook.  so always load from the result
if we're doing an update (the two revert cases).

still load on straight fork because we're not ready to ditch old github
which doesn't return the files in that case.

fixes #329