Moved load_module_package to authenticated ocaps
Quick request - the load_module_package is mostly used in authenticated scenarios for additional applications. This would be particularly useful for some internal application(hint - go and fish) which require authentication. This is a little critical so it would be very helpful if you could include this in this release.

Thanks,
SMS