suppress warnings and messages for add-on packages
fixes #934, very crudely

@s-u, is there a better way to do this - should we divert these messages to a
log or something?  In particular, we don't want to hear about the packages being
loaded, and we don't want to hear that `View` has been masked.

But I suppose if people are developing their own RCloud add-ons, they'll want to
know if something goes wrong, and this will keep them in the dark.