don't react to dragging html/images
fixes #1032: explicitly do not react to text/html

@sujitbehera27, this might interest you: we were *accidentally* disabling
the dragging of images within the page (such as the shadow edges of the cells,
or plots), which we don't want to handle as assets, or at least don't know how
to.

turns out that dt.items.length === 3 in this case, probably because it supports
multiple formats on the one file/selection.  so when we only supported dropping
one file, we were not reacting to images, but we need this fix now that we
support multiple files.

the other line enables the (+) for multiple-file drops; we weren't getting
feedback when dragging multiple files over the asset area.