Adds support for an authenticated redis instance
Submitted an issue and pull request against `rediscc` (s-u/rediscc#4), the underlying
redis library used in RCloud. This was changed and repushed as version
0.1-2.

Now users can specify a property in rcloud.conf for the password to
their redis instance.

These changes and a few more have been documented in the INSTALL.md and
README.md