set language mode for command prompt; massive command prompt refactor
fixes #1105
the command prompt code was a horrible, horrible mess, with the history
and command entry code within methods of the prompt. legacy of the push to 1.0

build history and entry in private methods the way we are doing elsewhere
add a bunch of private members
have sensible methods and propagation for the language

it makes a lot more sense, to me anyway.