do not hold onto `rcloud` object
@anatoliy, here's the cause of our disconnect errors.

The session and `rcloud` object get destroyed every time a new notebook is
loaded. And the RCloud extensions get reloaded. But the UI handlers that
the extensions created, stick around. If they refer to unqualified
`rcloud`, they will hold a reference to the object, with all of its dead
references to functions on the server.

So change them to explicitly use `window.rcloud` so that no reference is
held.