toward a cleaner lazy load
we don't need to reselect in tree_open anymore, because we are either
opening a tree by click (selection isn't in there), or we've proactively
lazy-loaded before we updated the model

there was never any reason to delay changing the current node when
loading its children, afaict - it is already created and save to modify

keep track of lazy load via a global map, not in the tree nodes, because
the tree nodes seem to be recreated from data

this still doesn't deal with friends right. in the last revision, we
were lazy loading them when opening friends or alls, which worked but
it's kind of inefficient and doesn't make sense now that we are
proactively lazy loading before updating the model.

(yes that's a little joke. it's still lazy loading but it should be
driven from the model, even though the model and view are so tangled.)

for #2144