mechanical refactoring
some clearly duplicated code here, and i need is_mine to be checked in
only one place for #831. simplify this without thinking about it too much.

@s-u, it's pretty amusing that the new function fork_and_name_notebook
switches between the self-fork workaround, and a case that says "don't
we wish we could fork a repo and immediately rename it but no, we run
into a github race condition"