put two of our github workarounds on one flag
for #831

this probably needs to be rewritten further - it just happens that the
self-fork workaround, and a place where we had to skip renaming a fork
because of the github gist api race condition, are in one place where i
can check a flag with a really stupid name.

the flag should probably be reversed, but again this is a first draft.

are there more places where we actually disable functionality because of
the race condition?