Improve notebook output autoscroll functionality
Introduce results buffering and process the results in batches. This removes fleakyness which was caused by race conditions between: R producing the results and pushing htmlwidgets contents, autoscroll feature and user scroll actions.

Thanks to this only scrolls up the screen need to be detected and cosidered as trigger for disabling autoscrolling.

The jquery.scrollto library got updated to alternative implementation which is actively maintained, allows for stopping scrolling animation on user scroll.

FIX #2522, #2551