Turn off detailed diff generation for large files.
In case of larger files (greater than 0.25MB) generating a diff resulted in a long execution of JS and/or crashing a web browser.

If a file is greater than 0.25MB the diff representing a replacement of whole file contents is generated to avoid the above issues.

This change also improves the performance of generating diffs for new and deleted cells/assets as the diff library isn't unnecessarily executed for such files.